requirements.txt
pydub
urllib3
deepgram-sdk
word2number
openai
langchain
nltk
language_tool_python